[rhi] More of the conversion from shaderpack::PipelineData to PipelineCreateInfo
