+++ src/renderer/pipeline_storage.cpp
@@ -95,93 +95,6 @@ namespace nova::renderer {
             });
     }
 
-    rhi::VertexFieldFormat to_rhi_vertex_format(const SPIRType& spirv_type) {
-        switch(spirv_type.basetype) {
-            case SPIRType::UInt:
-                return rhi::VertexFieldFormat::Uint;
-
-            case SPIRType::Float: {
-                switch(spirv_type.vecsize) {
-                    case 2:
-                        return rhi::VertexFieldFormat::Float2;
-
-                    case 3:
-                        return rhi::VertexFieldFormat::Float3;
-
-                    case 4:
-                        return rhi::VertexFieldFormat::Float4;
-
-                    default:
-                        logger(rx::log::level::k_error, "Nova does not support float fields with %u vector elements", spirv_type.vecsize);
-                        return rhi::VertexFieldFormat::Invalid;
-                }
-            };
-
-            case SPIRType::Unknown:
-                [[fallthrough]];
-            case SPIRType::Void:
-                [[fallthrough]];
-            case SPIRType::Boolean:
-                [[fallthrough]];
-            case SPIRType::SByte:
-                [[fallthrough]];
-            case SPIRType::UByte:
-                [[fallthrough]];
-            case SPIRType::Short:
-                [[fallthrough]];
-            case SPIRType::UShort:
-                [[fallthrough]];
-            case SPIRType::Int:
-                [[fallthrough]];
-            case SPIRType::Int64:
-                [[fallthrough]];
-            case SPIRType::UInt64:
-                [[fallthrough]];
-            case SPIRType::AtomicCounter:
-                [[fallthrough]];
-            case SPIRType::Half:
-                [[fallthrough]];
-            case SPIRType::Double:
-                [[fallthrough]];
-            case SPIRType::Struct:
-                [[fallthrough]];
-            case SPIRType::Image:
-                [[fallthrough]];
-            case SPIRType::SampledImage:
-                [[fallthrough]];
-            case SPIRType::Sampler:
-                [[fallthrough]];
-            case SPIRType::AccelerationStructureNV:
-                [[fallthrough]];
-            case SPIRType::ControlPointArray:
-                [[fallthrough]];
-            case SPIRType::Char:
-                [[fallthrough]];
-            default:
-                logger(rx::log::level::k_error, "Nova does not support vertex fields of type %u", spirv_type.basetype);
-        }
-
-        return {};
-    }
-
-    rx::vector<rhi::VertexField> PipelineStorage::get_vertex_fields(const ShaderSource& vertex_shader) const {
-        const CompilerGLSL shader_compiler{vertex_shader.source.data(), vertex_shader.source.size()};
-
-        const auto& shader_vertex_fields = shader_compiler.get_shader_resources().stage_inputs;
-
-        rx::vector<rhi::VertexField> vertex_fields;
-        vertex_fields.reserve(shader_vertex_fields.size());
-
-        for(const auto& spirv_field : shader_vertex_fields) {
-            const auto& spirv_type = shader_compiler.get_type(spirv_field.base_type_id);
-            const auto format = to_rhi_vertex_format(spirv_type);
-
-            vertex_fields.emplace_back(spirv_field.name.c_str(), format);
-        }
-
-        return vertex_fields;
-    }
-
     void PipelineStorage::get_shader_module_descriptors(const rx::vector<uint32_t>& spirv,
                                                         const rhi::ShaderStage shader_stage,
                                                         rx::map<rx::string, rhi::ResourceBindingDescription>& bindings) {
